Project 1 - Craftipedia
This is the final version of my app for project 1.  This app provides
the framework that I will use and expand upon in the net two projects.
It loads a UITableView with data that I created in an array.  The data
from the array is populated into a custom cell class that I created.
The table is also searchable.

Future iterations of the project will have an external database, detail
view once you click on an item, perhaps an indexed list with categories
to break down the alphabet, menu and loading screens, and more.